+++ Yacht63FT/schema/event.hpp
@@ -19,7 +19,7 @@ namespace WarGrey::SCADA {
     void restore_event(WarGrey::SCADA::IPreparedStatement* stmt, WarGrey::SCADA::AlarmEvent& self);
     void create_event(WarGrey::SCADA::IDBSystem* dbc, bool if_not_exists = true);
     void insert_event(WarGrey::SCADA::IDBSystem* dbc, AlarmEvent* self, bool replace = false);
-    void insert_event(WarGrey::SCADA::IDBSystem* dbc, AlarmEvent* selves, size_t count, bool replace = false);
+    void insert_event(WarGrey::SCADA::IDBSystem* dbc, AlarmEvent* selves, int count, bool replace = false);
     std::list<AlarmEvent> select_event(WarGrey::SCADA::IDBSystem* dbc, unsigned int limit = 0, unsigned int offset = 0);
     void drop_event(WarGrey::SCADA::IDBSystem* dbc);
 }
